Move logback to an single module
