+++ core/build.gradle.kts
@@ -6,7 +6,7 @@ dependencies {
 	api("org.slf4j:slf4j-api")
 	implementation("org.slf4j:slf4j-simple")
 	implementation("org.slf4j:jul-to-slf4j")
-	implementation("org.slf4j:log4j-to-slf4j")
+	implementation("org.apache.logging.log4j:log4j-to-slf4j")
 	testImplementation("org.springframework:spring-test")
 	testImplementation("org.junit.jupiter:junit-jupiter-api")
 	testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine")
